+++ backend-services/voting-app/src/main/java/org/cardano/foundation/voting/repository/VoteRepository.java
@@ -14,9 +14,6 @@ import java.util.Optional;
 public interface VoteRepository extends JpaRepository<Vote, String> {
 
     @Query("SELECT v FROM Vote v WHERE v.eventId = :eventId ORDER BY v.votedAtSlot, v.createdAt DESC")
-    List<Vote> findAllByEventId(@Param("eventId") String eventId);
-
-    @Query("SELECT v FROM Vote v WHERE v.eventId = :eventId ORDER BY v.votedAtSlot, v.createdAt DESC")
     List<CompactVote> findAllCompactVotesByEventId(@Param("eventId") String eventId);
 
     Optional<Vote> findByEventIdAndCategoryIdAndVoterStakingAddress(String eventId, String categoryId, String voterStakeAddress);