+++ ui/summit-2023/src/pages/Categories/Categories.tsx
@@ -17,7 +17,7 @@ import {
 import ViewModuleIcon from '@mui/icons-material/ViewModule';
 import ViewListIcon from '@mui/icons-material/ViewList';
 import { Fade } from '@mui/material';
-import './Categories.scss';
+import styles from './Categories.module.scss';
 import { CategoryContent } from './Category.types';
 import Grow from '@mui/material/Grow';
 import CATEGORY_IMAGES from '../../common/resources/data/categoryImages.json';
@@ -68,14 +68,26 @@ const Categories = () => {
   };
 
   return (
-    <>
+    <div
+      data-testid="categories-page"
+      className={styles.categories}
+    >
       <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: 20 }}>
         <Typography
-          className="categories-title"
+          className={styles.title}
           variant="h2"
+          fontSize={{
+            xs: '28px',
+            md: '32px',
+          }}
+          lineHeight={{
+            xs: '28px',
+            md: '32px',
+          }}
         >
           Categories
         </Typography>
+
         {!isMobile && (
           <div>
             <IconButton onClick={() => handleListView('grid')}>
@@ -96,13 +108,15 @@ const Categories = () => {
         {categories?.map((category, index) => (
           <Grid
             item
-            xs={!isMobile && listView === 'grid' ? 4 : 12}
+            xs={12}
+            sm={6}
+            md={!isMobile && listView === 'grid' ? 4 : 12}
             key={category.id}
           >
-            {listView === 'grid' || isMobile ? (
+            {!isMobile && listView === 'grid' ? (
               <Fade in={isVisible}>
                 <Card
-                  className="categories-card"
+                  className={styles.card}
                   sx={{
                     width: isMobile ? '100%' : '414px',
                   }}
@@ -309,7 +323,7 @@ const Categories = () => {
           </Grid>
         ))}
       </Grid>
-    </>
+    </div>
   );
 };