+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.2.33](https://github.com/cardano-foundation/cf-cardano-ballot/compare/v0.2.32...v0.2.33) (2023-09-13)
+
+
+### Bug Fixes
+
+* removed the unused component ([753da52](https://github.com/cardano-foundation/cf-cardano-ballot/commit/753da52098b98d2282205433c25a5b051f3b30e5))
+
+## [0.2.32](https://github.com/cardano-foundation/cf-cardano-ballot/compare/v0.2.31...v0.2.32) (2023-09-13)
+
+
+### Bug Fixes
+
+* leaderboard results preinitialized with 0. ([#208](https://github.com/cardano-foundation/cf-cardano-ballot/issues/208)) ([5faa9de](https://github.com/cardano-foundation/cf-cardano-ballot/commit/5faa9deca352a22a4cf8962e23373a411ad9d606))
+
 ## [0.2.31](https://github.com/cardano-foundation/cf-cardano-ballot/compare/v0.2.30...v0.2.31) (2023-09-13)